+++ src/main/frontend/src/App.js
@@ -1,14 +1,10 @@
 import logo from "./logo.svg";
 import "./App.css";
 import { BrowseRouter, Routes, Route } from "react-router-dom";
-import Layout from "./layout/Layout";
+import Layout from "./layouts/Layout";
 
 function App() {
-  return (
-    <Layout>
-      
-    </Layout>
-  ) 
+  return <Layout></Layout>;
 }
 
 export default App;
