--- src/main/frontend/src/layout/Footer.js
@@ -1,11 +1,0 @@
-import '../styles/layout/Footer.css'
-
-const Footer = () => {
-    return (
-        <footer className="footer">
-            <h2>푸터</h2>
-        </footer>
-    )
-}
-
-export default Footer
