--- src/main/frontend/src/layout/Header.js
@@ -1,11 +1,0 @@
-import '../styles/layout/Header.css'
-
-const Header = () => {
-    return (
-        <header className="header">
-            <h2>헤더</h2>
-        </header>
-    )
-}
-
-export default Header
