+++ src/main/frontend/src/layouts/Footer.js
@@ -1,0 +1,11 @@
+import "../styles/layouts/Footer.css";
+
+const Footer = () => {
+  return (
+    <footer className="footer">
+      <h2 style={{ marginTop: "0" }}>푸터</h2>
+    </footer>
+  );
+};
+
+export default Footer;
