+++ src/main/frontend/src/layouts/Layout.js
@@ -1,0 +1,18 @@
+import Header from "./Header";
+import Footer from "./Footer";
+import "../styles/layouts/Layout.css";
+import { Outlet } from "react-router";
+
+const Layout = () => {
+  return (
+    <div className="layout">
+      <Header />
+      <main className="main">
+        <Outlet />
+      </main>
+      <Footer />
+    </div>
+  );
+};
+
+export default Layout;
