--- src/main/frontend/src/styles/Layout.css
@@ -1,7 +1,0 @@
-.layout {
-    
-}
-
-.main {
-    margin-top: 500px;
-}
