--- src/main/frontend/src/styles/layout/Header.css
@@ -1,6 +1,0 @@
-.header {
-    position: fixed;
-    width: 100%;
-    top: 0;
-    background-color: rgb(187, 205, 110);
-}
