+++ src/main/frontend/src/styles/layouts/Footer.css
@@ -1,6 +1,5 @@
 .footer {
-    position: fixed;
     width: 100%;
-    margin-top: auto;
+    height: 70px;
     background-color: rgb(107, 83, 67);
 }
