+++ src/main/frontend/src/styles/layouts/Layout.css
@@ -1,0 +1,10 @@
+.layout {
+    display: flex; 
+    min-height: 100vh; 
+    flex-direction: column;
+}
+
+.main {
+    flex: 1;
+    height: fit-content;
+}
